Did some housekeeping: splitted large source files to smaller ones.
